rename author to info and make more generic
