--- stories/author.stories.js
@@ -1,31 +1,0 @@
-import { Author } from '@/components/author';
-
-const config = {
-  title: 'Components/Author',
-  component: Author,
-  parameters: {
-    layout: 'centered',
-    docs: {
-      toc: true,
-    },
-  },
-  tags: ['autodocs'],
-};
-
-export default config;
-
-export const Red = {
-  args: {
-    name: 'Example Name',
-    info: 'Enter a quote or other information about the author here',
-    color: 'red',
-  },
-};
-
-export const Yellow = {
-  args: {
-    name: 'Example Name',
-    info: 'Enter a quote or other information about the author here',
-    color: 'yellow',
-  },
-};
